// frontend/src/api-generated/models/linhKienPhanCungResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface LinhKienPhanCungResponse {
  id?: number;
  idTaiSanPhanCung?: number;
  tenTaiSanPhanCung?: string;
  maMauTaiSanPhanCung?: string;
  idNhaCungCap?: number;
  idDonVi?: number;
  soSerial?: string;
  giaMua?: number;
  thoiGianMua?: string;
  hanBaoHanhThang?: number;
  trangThaiKho?: string;
  viTriKho?: string;
  trangThai?: string;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}